Depends on WebKit r73697.

BUG=65619
TEST=browser_tests

Review URL: http://codereview.chromium.org/5695002

TBR=hans@chromium.org
Review URL: http://codereview.chromium.org/5717006

git-svn-id: svn://svn.chromium.org/chrome/branches/597/src@68993 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/common/indexed_db_key.cc
@@ -34,7 +34,7 @@ void IndexedDBKey::Set(const string16& string) {
   string_ = string;
 }
 
-void IndexedDBKey::Set(int32_t number) {
+void IndexedDBKey::Set(double number) {
   type_ = WebIDBKey::NumberType;
   number_ = number;
 }
